requirements.txt
pymupdf
python-docx
pydantic
pydantic-settings
langchain
langchain-core
langchain-deepseek
langchain-groq
langchain-openai
python-dotenv
python-multipart
fastapi
uvicorn
